Debounce move to global function under Hooks folder
