+++ src/Comps/search/Search.tsx
@@ -1,5 +1,6 @@
 import React, { ChangeEvent, useState ,useEffect} from 'react';
 import { SearchIcon } from "../../Icons";
+import useDebounce from '../../Hooks/useDebounce'
 import {
   SearchFormContainer,
   SearchButton,
@@ -9,19 +10,6 @@ import {
 
 // import DropDown from '../dropdown/DropDown';
 
-function useDebounce<T>(value: T, delay?: number): T {
-  const [debouncedValue, setDebouncedValue] = useState<T>(value)
-
-  useEffect(() => {
-    const timer = setTimeout(() => setDebouncedValue(value), delay || 500)
-
-    return () => {
-      clearTimeout(timer)
-    }
-  }, [value, delay])
-
-  return debouncedValue
-}
 
 const Search = () => {
   const [input, setInput] = useState<string>("");
