+++ src/Hooks/useDebounce.tsx
@@ -1,0 +1,17 @@
+import React, { useState ,useEffect} from 'react';
+
+function useDebounce<T>(value: T, delay?: number): T {
+    const [debouncedValue, setDebouncedValue] = useState<T>(value)
+  
+    useEffect(() => {
+      const timer = setTimeout(() => setDebouncedValue(value), delay || 500)
+  
+      return () => {
+        clearTimeout(timer)
+      }
+    }, [value, delay])
+  
+    return debouncedValue
+  }
+
+export default useDebounce
